Dragon egg cost fix
Fixed a problem with dragon eggs costing nothing until a dragon egg is
first place.  The issue had to do with the value of the dragon egg not
processing properly until at least one dragon egg was instantiated.  Now
if no value for the cost has been initialized, it simply costs the
default value.