Fireball feature implementation
Added features to fireball, including gold cost, cooldown timer, and
increasing cost each time it is used.  All of these values can be
adjusted during balancing.  Currently the number of ticks until the
fireball is ready again is displayed in the corner of the screen.  This
can later be replaced with some other indicator if we eventually have
assets to devote to it.