Restored chain lightning effect
I put the chain lightening effect back onto the lightning dragon
turrets.  It now jumps to a few different enemies after impact, doing
subsequently less damage to each in turn.  The number of bounces can be
adjusted by changing the "bounce" assignment in the _DragonLightning.cs
file when it shoots.

Another addition was a unique countdown timer on the
_ProjectileLightning.cs file which causes the projectile to disappear
after twenty ticks, unless something else destroys it first.  This was
to deal with a bug which caused the lightning bolts to sometimes persist
after they should have disappeared.