Fixed error in lightning projectile
Someone added in a decay variable in an attempt to delete the projectile
when a certain number of "ticks" had passed. I commented out the code
for this as this decay variable was not initialized, so when the game
checked if (decay <= 0), it would always return true, and immediately
destroy the projectile.
I added in a timeAlive variable which is increased by delta time each
update, and once the projectile has been alive for 3 seconds it is
deleted (for the rare times the shot misses the target)
Lightning dragons are now fully functioning once more, though I have
noticed the damage is significantly underpowered.