Fireball refinement
I have added some functionality to the fireball cooldown.  It now
reduces the cooldown in synch with the system clock rather than per
tick, so it will no longer slow down when the game gets active.  It also
displays per second rather than per tick, and clearly indicates when the
fireball is ready to launch again.